+++ parsers/royal.py
@@ -5,6 +5,7 @@ import html
 import re
 from sharedutils import errlog
 from parse import appender 
+import datetime
 
 def main():
     for filename in os.listdir('source'):
@@ -19,7 +20,10 @@ def main():
                     title = html.unescape(entry['title'])
                     website = str(entry['url'])
                     description = html.unescape((re.sub(r'<[^>]*>', '',entry['text'])))
-                    appender(title, 'royal', description.replace('\n',''),website)
+                    date_str = entry['time']
+                    dt_object = datetime.datetime.strptime(date_str, "%Y-%B-%d").replace(hour=1, minute=2, second=3, microsecond=456789)
+                    published = dt_object.strftime("%Y-%m-%d %H:%M:%S.%f")
+                    appender(title, 'royal', description.replace('\n',''),website,published)
                 file.close()
         except:
             errlog('royal: ' + 'parsing fail')